Fix name collision between Antrea NetworkPolicy and K8s NetworkPolicy
NetworkPolicy in controlplane API group is the object that consumed by
antrea-agents. Both Antrea NetworkPolicy and K8s NetworkPolicy will be
converted to it. Currently, the namespace and name of the original
NetworkPolicy are copied to the controlplane NetworkPolicy and
<Namespace>/<Name> is used as the key func. Therefore, one K8s
NetworkPolicy may overwrite the controlplane NetworkPolicy mapping to a
Antrea NetworkPolicy that has the same namespace and name.

This patch changes to use the UID of the original NetworkPolicy as the
name and UID of the controlplane NetworkPolicy and keeps the namespace,
name, UID, and type of the original NetworkPolicy in a new field
"SourceRef". Besides, the controlplane NetworkPolicy is changed to
cluster scoped like AppliedToGroups and AddressGroups.