Add Tier Mapping RBAC
This PR adds the following:
  - TierEntitlement and TierEntitlementBinding CRDs
  - YAMLs and OpenAPI validation
  - Handlers for authorization
TierEntitlement CRD is responsible to associate a Tier with either
"read" or "edit" permissions for the Tier's reference in any
Antrea-native policies.
TierEntitlementBinding CRD binds a TierEntitlement to User/Groups
in the cluster to authorize users to edit/read a Tier in
Antrea-native policies.

This feature can be enabled by setting the hidden config option
"EnterpriseAntrea: true" in antrea-controller.conf.