Fix Egress IP scheduling
PR #4593 introduced maxEgressIPsPerNode to limit the number of Egress IPs
that can be assigned to a Node. However, it used the EgressInformer cache
to check whether a Node can accommodate new Egress IPs and did the
calculation for different Egresses concurrently, which may cause
inconsistent schedule results among agents. For instance:

When Nodes' capacity is 1 and two Egresses, e1 and e2, are created
concurrently, different agents may process them in different orders, with
different contexts:

- agent a1 may process Egress e1 first and assign it to Node n1; it then
  processes Egress e2 and think it should be assigned to Node n2 by agent
  a2 because n1 is out of space.
- agent a2 may process Egress e1 and e2 faster, before any of their
  status is updated in Egress API, and would think both Egresses should
  be assigned to Node n1 by agent a1.

As a result, Egress e2 will be left unassigned.

To fix the problem, the Egress IP scheduling should be deterministic
accross agents and time. This patch adds an egressIPScheduler, which
takes the spec of Egress and ExternalIPPool and the state of memberlist
cluster as inputs, generates scheduling results deterministically.

According to the benchmark test, scheduling 1,000 Egresses among 1,000
Nodes once takes less than 3ms.

Signed-off-by: Quan Tian <qtian@vmware.com>