RegisterPacketInHandler should not use empty interface as parameter
The method is not supposed to handle values of unknown type, using empty
interface is obscure and unnecessary.

It was written in the current way to avoid an import cycle:

antrea.io/antrea/pkg/agent/openflow
-> antrea.io/antrea/pkg/agent/openflow/testing
--> antrea.io/antrea/pkg/agent/openflow

Normally the "testing" package is supposed to provide a mock
implementation of the interface defined in its parent package to clients
of the interface, and the parent package itself wouldn't import its own
testing package. However, the openflow Client makes an abstration of
OpenFlow operations via the "OFEntryOperations" interface, which refers
to itself in production code and is replaced by a mock implementation of
this interface in unit test, making the parent package import its
testing package.

This patch extracts pure OpenFlow operations to another package to
resolve the import cycle, as well as separating business specific and
operation generic code more clearly.

Signed-off-by: Quan Tian <qtian@vmware.com>